refactor content-process for sfari - 2
